not use babel-node in prod (+1 squashed commit) Squashed commits: [22c3410] rename index.js to server.js
